Improve function naming and split utils package
