+++ pkg/runner/shell.go
@@ -57,7 +57,7 @@ func demote(username, groupname string) (*syscall.SysProcAttr, error) {
 	}, nil
 }
 
-func runCmd(args []string, username, groupname string, env map[string]string, timeout int) (exitCode int, result string) {
+func runCmdWithOutput(args []string, username, groupname string, env map[string]string, timeout int) (exitCode int, result string) {
 	if env != nil {
 		defaultEnv := getDefaultEnv()
 		for key, value := range defaultEnv {
@@ -133,6 +133,19 @@ func runCmd(args []string, username, groupname string, env map[string]string, ti
 	}
 
 	return 0, string(output)
+}
+
+func RunCmd(command string, args ...string) int {
+	cmd := exec.Command(command, args...)
+
+	err := cmd.Run()
+	if err != nil {
+		if exitError, ok := err.(*exec.ExitError); ok {
+			return exitError.ExitCode()
+		}
+		return -1
+	}
+	return 0
 }
 
 // && and || operators are handled separately in handleShellCmd
