+++ pkg/scheduler/session.go
@@ -118,7 +118,7 @@ func (session *Session) do(req *http.Request, timeout time.Duration) ([]byte, in
 	req = req.WithContext(ctx)
 
 	req.Header.Set("Authorization", session.authorization)
-	req.Header.Set("User-Agent", utils.GetUserAgent())
+	req.Header.Set("User-Agent", utils.GetUserAgent("alpamon"))
 
 	if req.Method == http.MethodPost || req.Method == http.MethodPut || req.Method == http.MethodPatch {
 		req.Header.Set("Content-Type", "application/json")
@@ -201,7 +201,7 @@ func (session *Session) MultipartRequest(url string, body bytes.Buffer, contentT
 	req = req.WithContext(ctx)
 
 	req.Header.Set("Authorization", session.authorization)
-	req.Header.Set("User-Agent", utils.GetUserAgent())
+	req.Header.Set("User-Agent", utils.GetUserAgent("alpamon"))
 	req.Header.Set("Content-Type", contentType)
 
 	resp, err := session.Client.Do(req)
