+++ pkg/utils/utils.go
@@ -5,6 +5,9 @@ import (
 	"context"
 	"fmt"
 	"github.com/alpacanetworks/alpamon-go/pkg/version"
+	"github.com/google/go-github/github"
+	"github.com/rs/zerolog/log"
+	"github.com/shirou/gopsutil/v4/host"
 	"net/url"
 	"os"
 	"os/user"
@@ -12,13 +15,6 @@ import (
 	"runtime"
 	"strconv"
 	"strings"
-	"time"
-
-	"github.com/google/go-github/github"
-	"github.com/rs/zerolog/log"
-	"github.com/shirou/gopsutil/v4/disk"
-	"github.com/shirou/gopsutil/v4/host"
-	"github.com/shirou/gopsutil/v4/net"
 )
 
 var (
@@ -116,51 +112,6 @@ func ConvertGroupIds(groupIds []string) []uint32 {
 	return gids
 }
 
-func CalculateNetworkBps(current net.IOCountersStat, last net.IOCountersStat, interval time.Duration) (inputBps float64, outputBps float64) {
-	if interval == 0 {
-		return 0, 0
-	}
-
-	inputBytesDiff := float64(current.BytesRecv - last.BytesRecv)
-	outputBytesDiff := float64(current.BytesSent - last.BytesSent)
-	seconds := interval.Seconds()
-
-	inputBps = (inputBytesDiff * 8) / seconds
-	outputBps = (outputBytesDiff * 8) / seconds
-
-	return inputBps, outputBps
-}
-
-func CalculateNetworkPps(current net.IOCountersStat, last net.IOCountersStat, interval time.Duration) (inputPps float64, outputPps float64) {
-	if interval == 0 {
-		return 0, 0
-	}
-
-	inputPktsDiff := float64(current.PacketsRecv - last.PacketsRecv)
-	outputPktsDiff := float64(current.PacketsSent - last.PacketsSent)
-	seconds := interval.Seconds()
-
-	inputPps = inputPktsDiff / seconds
-	outputPps = outputPktsDiff / seconds
-
-	return inputPps, outputPps
-}
-
-func CalculateDiskIOBps(current disk.IOCountersStat, last disk.IOCountersStat, interval time.Duration) (readBps float64, writeBps float64) {
-	if interval == 0 {
-		return 0, 0
-	}
-
-	readBytesDiff := float64(current.ReadBytes - last.ReadBytes)
-	writeBytesDiff := float64(current.WriteBytes - last.WriteBytes)
-	seconds := interval.Seconds()
-
-	readBps = readBytesDiff / seconds
-	writeBps = writeBytesDiff / seconds
-
-	return readBps, writeBps
-}
-
 func Quote(s string) string {
 	if len(s) == 0 {
 		return "''"
@@ -204,6 +155,6 @@ func GetLatestVersion() string {
 	return release.GetTagName()
 }
 
-func GetUserAgent() string {
-	return fmt.Sprintf("%s/%s", "alpamon", version.Version)
+func GetUserAgent(name string) string {
+	return fmt.Sprintf("%s/%s", name, version.Version)
 }
